Always apply declaration merging
