+++ src/index.ts
@@ -1,6 +1,15 @@
 import fp from 'fastify-plugin'
 import { flashFactory } from './flash'
 
+declare module 'fastify' {
+  export interface FastifyRequest {
+    flash: ReturnType<typeof flashFactory>['request']
+  }
+  export interface FastifyReply {
+    flash: ReturnType<typeof flashFactory>['reply']
+  }
+}
+
 export = fp<{}>(
   function (fastify, opts, done) {
     const flash = flashFactory()
